+++ tests/teleconnections/test_mjo.py
@@ -11,6 +11,7 @@ approx_rel = APPROX_REL
 loglevel = LOGLEVEL
 
 @pytest.mark.diagnostics
+@pytest.mark.skip(reason="Temporarily disabled until MJO is integrated again.")
 def test_mjo(tmp_path):
     """
     Test that the MJO class works
@@ -20,7 +21,9 @@ def test_mjo(tmp_path):
         'exp': 'era5-hpz3',
         'source': 'monthly',
         'loglevel': loglevel,
-        'regrid': 'r200'
+        'regrid': 'r200',
+        'startdate': '2000-01-01',
+        'enddate': '2001-12-31'
     }
 
     mjo = MJO(**init_dict)